refactor(WinningResultView): 로또 당첨 내역을 출력하는 뷰 클래스를 생성
